+++ frontend/public/changelog.md
@@ -5,6 +5,7 @@
 - Menyederhanakan tampilan beranda
 - Memberikan pilihan pencarian kamus, tesaurus, dan glosarium di kotak cari
 - Memberikan judul unik untuk tiap halaman
+- Menampilkan kategori di halaman kamus
 
 ## 2026-02-14 Sabtu